add allocate_bom implementation in Allocatable, PatchMaker implementation just wraps that
